pyproject.toml
[tool.poetry]
name = "repl_python3_COURSE-WORK"
version = "0.1.0"
description = ""
authors = ["ojbgrsx <<>>"]

[tool.poetry.dependencies]
python = "^3.8"
google-api-python-client = "^2.30.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
